LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-controls-bar-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
.media-controls {
    position: absolute;
    top: 0;
    left: 0;
}
    
</style>
<script type="text/javascript">

description("Testing the <code>MacOSFullscreenMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSFullscreenMediaControls;

let style, bounds;
scheduler.frameDidFire = function()
{
    document.body.appendChild(mediaControls.element);

    style = window.getComputedStyle(mediaControls.controlsBar.element);
    bounds = mediaControls.controlsBar.element.getBoundingClientRect();
    
    shouldBeEqualToString("style.left", "50%");
    shouldBeEqualToString("style.bottom", "25px");
    shouldBeEqualToString("style.width", "468px");
    shouldBeEqualToString("style.height", "75px");
    shouldBeEqualToString("style.transform", "matrix(1, 0, 0, 1, -234, 0)");
    shouldBeEqualToString("style.borderRadius", "8px");

    shouldBe("bounds.left", "-84");
    shouldBe("bounds.top", "50");
    shouldBe("bounds.width", "468");
    shouldBe("bounds.height", "75");

    shouldBeEqualToRGBAColor("style.backgroundColor", "rgba(30, 30, 30, 0.45)");
    shouldBeEqualToString("style.webkitBackdropFilter", "saturate(1.8) blur(20px)");

    mediaControls.element.remove();

    debug("");
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
